requirements.txt
sqlalchemy==2.0.4
black==22.3.0
pytest==7.2.1
python-dotenv==0.20.0
freezegun==0.3.4
pytest-mock==3.10.0
